Interrupt: only use 'u32' as identifier type of IRQs

+++ include/gba/interrupt.h
@@ -77,6 +77,6 @@ INLINE void interrupt_toggle(u32 irq, bool enable) {
     }
 }
 
-extern void interrupt_set_isr(u8 irq, void (*isr)(void));
+extern void interrupt_set_isr(u32 irq, void (*isr)(void));
 
 #undef _INTERRUPT_IE
